+++ .gitignore
@@ -3,3 +3,4 @@
 /output_frames
 /output
 /lottie-rs
+/assets/output_frames
